+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Version 0.9.0
+
+- many functions got overloads for `Matcher` type propagation in less common scenarios;
+- `condition` function now accepts Parsers/Matchers with different value types, result value type is the union of the two;
+- added type tests for overloads using [tsd](https://github.com/SamVerschueren/tsd).
+
 ## Version 0.8.0
 
 - Targeting Node.js version 14 and ES2020;